+++ R/grid_helpers.R
@@ -19,6 +19,7 @@
 
 
 ## grid helpers
+## requires map.R
 
 .grid_basic_layout <- function(main = "",
   left = unit(4, "lines"),
@@ -143,79 +144,11 @@
 }
 
 
-
-### new version below uses grid.raster
-.grid_image_old <- function(x,
-  y,
-  z,
-  zlim,
-  col = grays(12),
-  name = "image",
-  gp = gpar()) {
-  if (is.matrix(x)) {
-    z <- x
-    x <- 1:ncol(z)
-    y <- 1:nrow(z)
-  }
-
-  if (missing(zlim))
-    zlim <- range(z, na.rm = TRUE)
-  else {
-    # fix data for limits
-    z[z < zlim[1]] <- NA
-    z[z > zlim[2]] <- NA
-  }
-
-  offset <- if (zlim[1] < 0)
-    - zlim[1]
-  else
-    0
-  range <- diff(zlim)
-
-  div <- 1 / length(col)
-
-  ## create a viewport
-  vp <- viewport(
-    xscale = c(0, (length(x) + 1)),
-    yscale = c((length(y) + 1), 0),
-    default.units = "native",
-    name = name
-  )
-  pushViewport(vp)
-
-  ## make sure we have a color for the maximal value (see floor +1)
-  col[length(col) + 1] <- col[length(col)]
-
-  ## the highest value is lightest color!
-  xs <- sapply(x, "rep.int", times = length(y))
-  grid.rect(
-    x = xs,
-    y = y,
-    1,
-    1,
-    gp = gpar(fill = col[floor((z + offset) / range / div) + 1], col = 0),
-    default.units = "native"
-  )
-
-  ## make border
-  gp_border       <- gp
-  gp_border$fill  <- "transparent"
-  grid.rect(
-    x = (length(x) + 1) / 2,
-    y = (length(y) + 1) / 2,
-    width = length(x),
-    height = length(y),
-    default.units = "native",
-    gp = gp_border
-  )
-
-  upViewport(1)
-}
-
 .grid_image <- function(x,
   zlim,
-  col = grays(12),
+  col = grDevices::gray.colors(12),
   prop = FALSE,
+  interpolate = FALSE,
   name = "image",
   gp = gpar()) {
   if (missing(zlim))
@@ -225,15 +158,6 @@
     x[x < zlim[1]] <- NA
     x[x > zlim[2]] <- NA
   }
-
-  offset <- if (zlim[1] < 0)
-    - zlim[1]
-  else
-    0
-  #range <- diff(zlim) ### not al plots have 0 as the minimum!
-  range <- zlim[2] + offset
-
-  div <- 1 / length(col)
 
   ## create a viewport
   if (!prop) {
@@ -260,7 +184,6 @@
       h <- 1
     }
     vp <- viewport(
-      #xscale = c(0,ncol(x)), yscale = c(nrow(x),0),
       xscale = c(0.5, ncol(x) + .5),
       yscale = c(nrow(x) + .5, 0.5),
       width = unit(w, "snpc"),
@@ -271,14 +194,11 @@
     pushViewport(vp)
   }
 
-  ## make sure we have a color for the maximal value (see floor +1)
-  col[length(col) + 1] <- col[length(col)]
+  x[] <- col[map_int(x, length(col))]
 
-  ## the highest value is lightest color!
-  x[] <- col[floor((x + offset) / range / div) + 1]
   grid.raster(
     x,
-    interpolate = FALSE,
+    interpolate = interpolate,
     default.units = "npc",
     width = 1,
     height = 1
@@ -328,61 +248,6 @@
   ## hopefully there is space outside for axes
   grid.xaxis()
   grid.text(xlab, y = unit(-3, "lines"))
-
-  upViewport(1)
-}
-
-# new colorkey uses grid.raster
-.grid_colorkey_old <- function(range,
-  col,
-  threshold = NULL,
-  name = "colorkey",
-  gp = gpar()) {
-  vp <- viewport(
-    xscale = range,
-    yscale = c(0, 1),
-    default.units = "native",
-    name = name
-  )
-  pushViewport(vp)
-
-  n <- length(col)
-  width <- diff(range) / n
-  xs <-
-    seq(range[1] + width / 2, range[2] - width / 2, length.out = n)
-
-  ## do not display the part above the threshold
-  col[xs > threshold] <- NA
-
-  ## col
-  gp_col      <- gp
-  gp_col$col  <- 0
-  gp_col$fill <- col
-  grid.rect(
-    x = xs,
-    y = 0,
-    width = width,
-    height = 1,
-    just = c("centre", "bottom"),
-    default.units = "native",
-    gp = gp_col
-  )
-
-
-  ## box
-  gp_border       <- gp
-  gp_border$fill  <- "transparent"
-  grid.rect(
-    x = 0,
-    y = 0,
-    width = 1,
-    height = 1,
-    just = c("left", "bottom"),
-    default.units = "npc",
-    gp = gp_border
-  )
-
-  grid.xaxis(gp = gp)
 
   upViewport(1)
 }