+++ R/map.R
@@ -1,5 +1,5 @@
 #######################################################################
-# Code to map between ranges for contunuous variables
+# Code to map between ranges for continuous variables
 # Copyright (C) 2011 Michael Hahsler
 #
 # This program is free software; you can redistribute it and/or modify
@@ -24,28 +24,36 @@ map <- function(x,
     if (any(is.na(from.range)))
         from.range <- range(x, na.rm = TRUE)
 
-    ## check if all values are the same
-    if (!diff(from.range))
-        return(matrix(
-            mean(range),
-            ncol = ncol(x),
-            nrow = nrow(x),
-            dimnames = dimnames(x)
-        ))
+    if (length(from.range) != 2L ||
+            from.range[1] > from.range[2])
+        stop('from.range needs to contain 2 numbers (upper <= lower bound).')
+    from.range_width <- from.range[2] - from.range[1]
+
+    if (length(range) != 2L)
+        stop('range needs to contain 2 numbers (upper and lower bound).')
+    range_width <- range[2] - range[1]
+
+    ## if all values are the same and no from.range is given, then return the average range
+    if (from.range_width == 0) {
+        x[] <- mean(range)
+        return(x)
+    }
 
     ## map to [0,1]
-    x <- (x - from.range[1])
-    x <- x / diff(from.range)
-    ## handle single values
-    if (diff(from.range) == 0)
-        x <- 0
+    x <- (x - from.range[1]) / from.range_width
 
     ## map from [0,1] to [range]
-    if (range[1] > range[2])
-        x <- 1 - x
-    x <- x * (abs(diff(range))) + min(range)
-
-    x[x < min(range) | x > max(range)] <- NA
+    x <- x * range_width + range[1]
 
     x
 }
+
+map_int <- function(x,
+    range = c(1L, 100L),
+    from.range = NA) {
+
+    if (length(range) == 1L) range <- c(1L, range)
+    as.integer(map(x, c(range[1], range[2]), from.range))
+}
+
+